Day 10, both parts after long suffering.
Took many hours to eventually stumble upon correct pages on wikipedia.
I initially attempted to solve this using flood fill like in Day 18 / 2022,
but that proved to be difficult with the pipes to squeeze through. I then
tried to solve it by casting rays and counting the number of intersections,
but didn't managed to get a working solution that way either. I then tried
to increase the resolution of the grid by replacing each grid item with 3x3
but that would've required a lot of refactoring. At that time I found
shoelace formula, and a bit later realized to combine it with Pick's theorem.